fix: add updated lake-manifest.json
I assumed that lake-manifest.json would behave like Cargo.lock,
and that it should not be pushed to GitHub in library code. However,
the Lean CI Action seems to expect it to be there, and it's also
not included in the pregenerated `.gitignore`, so I'm adding the
file back again, but updated this time.